requirements.txt
praw>=7.7.1
asyncio-throttle>=1.0.2
asyncpraw>=7.7.1
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=11.0
aiohttp>=3.8.0
schedule>=1.2.0